requirements.txt
python-binance==1.0.19
openai==1.3.0
pinecone-client==2.2.4
pandas>=2.1.3,<2.3.0
numpy>=1.24.3,<1.26.0
ta==0.10.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
redis==5.0.1
upstash-redis==1.0.0
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
neon-connector==0.0.1
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.35.2
yfinance==0.2.22
ccxt==4.1.77
schedule==1.2.0
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2